Use pushd / popd so pwd is preserved in shell
Using `cd` has the side effect of moving the pwd of the pane being used
for the test to the target package. Here we switch to using `pushd` and
`popd` so the test command has no side effects on the working directory.

I believe `pushd` and `popd` are present in all modern shells, so this
should present no interop problems.